Funkcijų aprašymus perkelti į atskirus .cpp failus.
